[FEAT] 크롤링 스케줄러 추가

+++ src/main/java/com/catspot/crawler/StudyPlace.java
@@ -12,7 +12,7 @@ import lombok.*;
 @Builder
 public class StudyPlace {
     @Id
-    private Integer placeIdx;
+    private Long placeIdx;
     private String placeName;
     private String url;
     private Integer allSeats;
